requirements.txt
# pip freeze --local >> requirements.txt
#
Django==1.6.8
MySQL-python==1.2.5
Pillow==2.6.1
South==1.0.1
Whoosh==2.6.0
argparse==1.2.2
beautifulsoup4==4.3.2
cmsplugin-poll==0.4
cmsplugin-zinnia==0.6
dj-database-url==0.3.0
django-app-namespace-template-loader==0.2
django-appconf==0.6
django-blog-zinnia==0.14.3
django-classy-tags==0.5.1
django-cms==3.0.7
django-cms-wunderground==1.0.0
django-compressor==1.4
django-contrib-comments==1.5
django-debug-toolbar==1.2.2
django-haystack==2.3.1
django-mptt==0.6.0
django-reversion==1.8.5
django-sekizai==0.7
django-tagging==0.3.4
django-xmlrpc==0.1.5
djangocms-admin-style==0.2.4
djangocms-column==1.5
djangocms-file==0.1
djangocms-flash==0.1
djangocms-googlemap==0.2
djangocms-inherit==0.1
djangocms-installer==0.6.0
djangocms-link==1.5
djangocms-picture==0.1
djangocms-style==1.5
djangocms-teaser==0.1
djangocms-text-ckeditor==2.4.1
djangocms-video==0.1
html5lib==0.999
pyparsing==2.0.3
python-memcached==1.53
pytz==2014.10
requests==2.5.0
six==1.8.0
sqlparse==0.1.14
zinnia-theme-bootstrap==0.3.1
zinnia-threaded-comments==0.3
zinnia-wysiwyg-wymeditor==1.1
